Updates to Function Runtime Settings UI * Don't show current running version * Don't show extension bundles version * Disable Runtime Version dropdown if a custom version is set
