+++ client-react/src/pages/app/app-settings/Sections/FunctionRuntimeSettingsPivot.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import { useTranslation } from 'react-i18next';
 import { isEqual } from 'lodash-es';
 import { style } from 'typestyle';
@@ -7,22 +7,40 @@ import { findFormAppSettingValue } from '../AppSettingsFormData';
 import DailyUsageQuota from '../FunctionRuntimeSettings/DailyUsageQuota';
 import HostJsonConfiguration from '../FunctionRuntimeSettings/HostJsonConfiguration';
 import RuntimeVersion from '../FunctionRuntimeSettings/RuntimeVersion';
+import RuntimeVersionBanner from '../FunctionRuntimeSettings/RuntimeVersionBanner';
 import RuntimeScaleMonitoring from '../FunctionRuntimeSettings/RuntimeScaleMonitoring';
 import { CommonConstants } from '../../../../utils/CommonConstants';
 import { ScenarioIds } from '../../../../utils/scenario-checker/scenario-ids';
 import { ScenarioService } from '../../../../utils/scenario-checker/scenario.service';
+import { PermissionsContext } from '../Contexts';
+import { ThemeContext } from '../../../../ThemeContext';
+import { MessageBar, MessageBarType } from 'office-ui-fabric-react';
+import { messageBannerStyle } from '../AppSettings.styles';
 
 const tabContainerStyle = style({
   marginTop: '15px',
 });
 
 const FunctionRuntimeSettingsPivot: React.FC<AppSettingsFormProps> = props => {
+  const { app_write, editable } = useContext(PermissionsContext);
+  const theme = useContext(ThemeContext);
   const { t } = useTranslation();
   const scenarioChecker = new ScenarioService(t);
   const site = props.initialValues.site;
 
   return (
     <div id="function-runtime-settings" className={tabContainerStyle}>
+      {(!app_write || !editable) && (
+        <MessageBar
+          id="function-runtime-settings-rbac-message"
+          isMultiline={true}
+          className={messageBannerStyle(theme, MessageBarType.warning)}
+          messageBarType={MessageBarType.warning}>
+          {t('readWritePermissionsRequired')}
+        </MessageBar>
+      )}
+      <RuntimeVersionBanner {...props} />
+
       {site.properties.state && site.properties.state.toLowerCase() === CommonConstants.SiteStates.running && <RuntimeVersion {...props} />}
 
       {scenarioChecker.checkScenario(ScenarioIds.runtimeScaleMonitoringSupported, { site }).status === 'enabled' && (
